+++ src/main/java/com/nineone/nocm/controller/websocket/MessageController.java
@@ -40,8 +40,12 @@ public class MessageController {
 		// 메세지 전송하기 전에 DB에 메세지 저장하는 로직이 있어야함. and 유저정보 가져오는 로직..?
 		message.setSend_date(messageService.makeDate());
 		message.setStr_send_date(messageService.makeStrDate(message.getSend_date()));
-		messageService.insertMessage(message);
-		messagingTemplate.convertAndSend("/sub/chat/room/"+message.getChannel_id(), message); 
+		if(messageService.insertMessage(message) > 0) {
+			messagingTemplate.convertAndSend("/sub/chat/room/"+message.getChannel_id(), message);
+		}else {
+			messagingTemplate.convertAndSend("/sub/"+message.getSender(), message);
+		}
+		 
 	}
